Alterado o tipo de altura para Integer e atualizados alguns elementos…

+++ db/migrate/20201012194759_add_fields_to_pacientes.rb
@@ -2,7 +2,7 @@ class AddFieldsToPacientes < ActiveRecord::Migration[6.0]
   def change
     add_column :pacientes, :nome, :string
     add_column :pacientes, :cpf, :integer
-    add_column :pacientes, :altura, :float
+    add_column :pacientes, :altura, :integer
     add_column :pacientes, :data_nascimento, :datetime
   end
 end
